projects/vdk-core/plugins/vdk-postgres/requirements.txt
vdk-core
psycopg2-binary

# testing requirements
click
vdk-test-utils
pytest-docker
docker-compose
